requirements.txt
numpy==2.3.5
opencv_contrib_python==4.10.0.84
opencv_python==4.11.0.86
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.3.3
Pillow==12.0.0
rfdetr==1.2.1
torch==2.7.0
torch==2.7.1
tqdm==4.67.1
ultralytics==8.3.145
ultralytics==8.3.160
